fix: fix holy site waypoint marking with coordinate system fixes
Fixes critical bugs preventing waypoint markers from displaying correctly on the map:

- Fix integer overflow in HolySiteData.GetCenter() by casting to long before multiplication
  (coordinates like 512035 * volume would overflow int32, causing incorrect negative values)
- Convert absolute world coordinates to spawn-relative coordinates for waypoint commands
  (Vintage Story's /waypoint addat expects relative coords, not absolute)
- Change waypoint icon from circle to star1 for better visibility
- Use named .NET colors (red, green, etc.) instead of hex codes for compatibility
- Add debug logging for coordinate conversion troubleshooting
- Preserve holy site names with spaces in waypoint titles

The waypoint system now correctly creates map markers at holy site locations with
colored star icons based on deity domain.

+++ DivineAscension/Data/HolySiteData.cs
@@ -238,9 +238,11 @@ public class HolySiteData
             int centerY = (area.Y1 + area.Y2) / 2;
             int centerZ = (area.Z1 + area.Z2) / 2;
 
-            totalX += centerX * volume;
-            totalY += centerY * volume;
-            totalZ += centerZ * volume;
+            // Cast to long before multiplication to prevent integer overflow
+            // Block coordinates can be very large (500k+), causing overflow when multiplied by volume
+            totalX += (long)centerX * volume;
+            totalY += (long)centerY * volume;
+            totalZ += (long)centerZ * volume;
             totalVolume += volume;
         }
